+++ services/git-service/src/app-key/GET_APP_GEY.gql.ts
@@ -1,0 +1,17 @@
+import { gql } from 'apollo-server-core';
+
+const GET_APP_KEY = gql`
+  query {
+    id
+    user {
+      id
+      email
+      name
+      role
+    }
+    key
+    name
+  }
+`;
+
+export default GET_APP_KEY;
