+++ services/git-service/src/app-key/app-key.resolver.ts
@@ -26,28 +26,3 @@ export class AppKeyResolver {
     return await this.prisma.mutation.updateAppKey(args, info);
   }
 }
-
-// @Resolver()
-// export class UsersResolver {
-//   constructor(private readonly prisma: PrismaService) {}
-
-//   @Query('users')
-//   async getUsers(@Args() args, @Info() info): Promise<User[]> {
-//     return await this.prisma.query.users(args, info);
-//   }
-
-//   @Query('user')
-//   async getUser(@Args() args, @Info() info): Promise<User> {
-//     return await this.prisma.query.user(args, info);
-//   }
-
-//   @Mutation('createUser')
-//   async createUser(@Args() args, @Info() info): Promise<User> {
-//     return await this.prisma.mutation.createUser(args, info);
-//   }
-
-//   @Mutation('updateUser')
-//   async updateUser(@Args() args, @Info() info): Promise<User> {
-//     return await this.prisma.mutation.updateUser(args, info);
-//   }
-// }
