+++ services/git-service/src/auth/http.strategy.ts
@@ -14,6 +14,7 @@ export class HttpStrategy extends PassportStrategy(Strategy) {
     if (!user) {
       throw new UnauthorizedException();
     }
+
     return user;
   }
 }
