+++ services/git-service/src/config/index.ts
@@ -1,0 +1,7 @@
+const GITHUB_GRAPHQL_URL = 'https://api.github.com/graphql';
+
+const config = {
+  GITHUB_GRAPHQL_URL,
+};
+
+export default config;
