+++ services/git-service/src/repositories/GET_REPOSITORIES.gql.ts
@@ -1,0 +1,13 @@
+import { gql } from 'apollo-server-core';
+
+const GET_REPOSITORIES = gql`
+  query {
+    id
+    idExternal
+    name
+    owner
+    isTracked
+  }
+`;
+
+export default GET_REPOSITORIES;
