+++ services/git-service/src/repositories/repositories.module.ts
@@ -1,10 +1,10 @@
+import { forwardRef, Module } from '@nestjs/common';
+import { PrismaModule } from '../prisma/prisma.module';
 import { AppKeyModule } from './../app-key/app-key.module';
 import { GitClientModule } from './../git-client/git-client.module';
-import { Module, forwardRef } from '@nestjs/common';
-import { RepositoriesService } from './repositories.service';
 import { RepositoriesController } from './repositories.controller';
 import { RepositoriesResolver } from './repositories.resolver';
-import { PrismaModule } from '../prisma/prisma.module';
+import { RepositoriesService } from './repositories.service';
 
 @Module({
   imports: [PrismaModule, forwardRef(() => GitClientModule), forwardRef(() => AppKeyModule)],
