+++ services/git-service/src/users/users.controller.ts
@@ -1,43 +1,26 @@
-import { AppKeyService } from './../app-key/app-key.service';
-import { UsersResolver } from './users.resolver';
-import { Controller, Get, Put, Body, Post, Req, Param } from '@nestjs/common';
-// import { gql } from 'apollo-server-core';
+import { Body, Controller, Get, Param, Post, Put, Req } from '@nestjs/common';
 import { AuthService } from '../auth/auth.service';
+import { AppKeyService } from './../app-key/app-key.service';
 import { GET_USERS } from './GET_USERS.gql';
+import { UsersResolver } from './users.resolver';
+import { UsersService } from './users.service';
 
 @Controller('users')
 export class UsersController {
   constructor(
-    private readonly usersResolver: UsersResolver,
+    private readonly usersService: UsersService,
     private readonly authService: AuthService,
     private readonly appKeyService: AppKeyService,
   ) {}
 
   @Get()
   async findAll() {
-    return this.usersResolver.getUsers({}, GET_USERS);
+    return await this.usersService.getAll();
   }
 
   @Get(['whoami/:id', ':id/data'])
   async whoAmI(@Req() request, @Param('id') id) {
-    const whoami = await this.usersResolver.getUser(
-      {
-        where: {
-          id,
-        },
-      },
-      GET_USERS,
-    );
-
-    const result = {
-      email: whoami.email,
-      id: whoami.id,
-      hasAppKey: whoami.keys && whoami.keys.length > 0 ? true : false,
-      name: whoami.name,
-      gitLogin: whoami.gitLogin,
-    };
-
-    return result;
+    return await this.usersService.getById(id);
   }
 
   @Put('/add-key')
@@ -64,34 +47,7 @@ export class UsersController {
     @Body('appKey') appKey = null,
   ) {
     console.log('Trying to retrieve existing user ' + email + ' ' + password + ' ' + appKey);
-    const existingUser = await this.usersResolver.getUser(
-      {
-        where: {
-          email,
-        },
-      },
-      GET_USERS,
-    );
-    console.log('Retrieved user');
 
-    if (!existingUser) {
-      console.log('Trying to create user');
-      const hashedPassword = await this.authService.encryptPassword(password);
-      const user = await this.usersResolver.createUser(
-        {
-          data: {
-            email,
-            password: hashedPassword,
-          },
-        },
-        GET_USERS,
-      );
-      console.log('returining new user');
-
-      return user;
-    }
-    console.log('returning existing user');
-
-    return existingUser;
+    return await this.usersService.upsertRetrieval(email, password, appKey);
   }
 }
