+++ services/webapp-service/src/common/HttpClient/index.ts
@@ -1,0 +1,9 @@
+import AbstractHttpClient from './HttpClient.abstract';
+import { HttpClient } from './HttpClient';
+
+export * from './repositories.client';
+export * from './users.client';
+
+export { AbstractHttpClient };
+
+export default new HttpClient();
