+++ services/webapp-service/src/common/HttpClient/repositories.client.ts
@@ -1,14 +1,19 @@
 import AbstractHttpClient from './HttpClient.abstract';
-import HttpClient from '@/common/HttpClient';
-import { AxiosInstance } from 'axios';
 
+/**
+ * Repositories-specific domain of the HttpClient
+ */
 export class RepositoriesClient {
   constructor(private readonly client: AbstractHttpClient) {}
 
   public getRepositories() {
-    console.log('get repositories');
+    // console.log('get repositories');
   }
 
+  /**
+   * Load all currently tracked repositories and the trackable repositories of the given owner
+   * @param owner
+   */
   public async loadSelectableRepositories(owner?: string) {
     if (owner == null) {
       owner = this.client.user.gitLogin;
@@ -26,6 +31,10 @@ export class RepositoriesClient {
     }
   }
 
+  /**
+   * Toggle the `isTracked` state of a repository
+   * @param repositoryData
+   */
   public async toggleRepositoryTracking(repositoryData: any) {
     await this.client.post(
       `${AbstractHttpClient.apiUrl}/repositories/${this.client.user.id}/toggle-tracking/${repositoryData.idExternal}`,
