+++ services/webapp-service/src/common/HttpClient/users.client.ts
@@ -1,8 +1,14 @@
 import AbstractHttpClient from './HttpClient.abstract';
 
+/**
+ * Users-specific domain of the HttpClient
+ */
 export class UsersClient {
   constructor(private readonly client: AbstractHttpClient) {}
 
+  /**
+   * Load User Data for a given user
+   */
   async loadUserData() {
     const result = await this.client.get(
       `${AbstractHttpClient.apiUrl}/users/${this.client.user.id}/data`
@@ -11,6 +17,9 @@ export class UsersClient {
     return result.data;
   }
 
+  /**
+   * Add an AppKey and attach it to the given user
+   */
   async addAppKey(username: string, repoName: string, key: string, user: string) {
     await this.client.put(`${AbstractHttpClient.apiUrl}/users/add-key`, {
       username,
@@ -20,6 +29,11 @@ export class UsersClient {
     });
   }
 
+  /**
+   * Validate an App Key
+   * @param username
+   * @param key
+   */
   async validateAppKey(username: string, key: string) {
     try {
       const isValid = await this.client.post(`${AbstractHttpClient.apiUrl}/users/validate-key`, {
