+++ services/webapp-service/src/config/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Base Api URL for absolute requests
+ */
+const API_URL = process.env.VUE_APP_GIT_SERVICE_HOST;
+
+const config = {
+  API_URL
+};
+
+export default config;
