+++ services/webapp-service/src/router/RouteManager/IRouteManagerContext.interface.ts
@@ -1,0 +1,6 @@
+/**
+ * Extensible context in the RouteManager for defining contextual parameters around route customizations
+ */
+export interface IRouteManagerContext {
+  hasJwt?: boolean;
+}
