+++ services/webapp-service/src/router/routes/IRoute.interface.ts
@@ -1,0 +1,5 @@
+import { RouteConfig } from 'vue-router/types/router';
+
+export interface IRoute extends RouteConfig {
+  displayName: string;
+}
