+++ services/webapp-service/src/router/routes/home.ts
@@ -1,0 +1,12 @@
+import { IRoute } from './IRoute.interface';
+import Home from '@/views/Home.vue';
+
+export const HOME: IRoute = {
+  path: '/',
+  name: 'home',
+  displayName: 'Home',
+  component: Home,
+  meta: {
+    requiresAuth: true
+  }
+};
