+++ services/webapp-service/src/router/routes/index.ts
@@ -1,0 +1,6 @@
+import * as routes from './routes';
+
+export * from './routes';
+export * from './IRoute.interface';
+
+export default routes.ROUTES;
