+++ services/webapp-service/src/router/routes/routes.ts
@@ -1,0 +1,8 @@
+import Home from '@/views/Home.vue';
+import { LOGIN } from './login';
+import { HOME } from './home';
+import { MY_REPOSITORIES } from './myRepositories';
+
+const ROUTES = [LOGIN, HOME, MY_REPOSITORIES];
+
+export { LOGIN, HOME, MY_REPOSITORIES, ROUTES };
