+++ services/webapp-service/src/store/aspects/app/IAppState.interface.ts
@@ -1,0 +1,3 @@
+export interface IAppState {
+  navExpanded: boolean;
+}
