+++ services/webapp-service/src/store/aspects/app/index.ts
@@ -1,0 +1,5 @@
+import AppStateModule from './app.store';
+
+export * from './IAppState.interface';
+
+export default AppStateModule;
