+++ services/webapp-service/src/store/aspects/repositories/IRepositoriesState.interface.ts
@@ -1,0 +1,3 @@
+export interface IRepositoriesState {
+  repositories: any[];
+}
