+++ services/webapp-service/src/store/aspects/repositories/index.ts
@@ -1,0 +1,5 @@
+import RepositoriesStateModule from './repositories.store';
+
+export * from './IRepositoriesState.interface';
+
+export default RepositoriesStateModule;
