+++ services/webapp-service/src/store/index.ts
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
-import { IAppState } from '@/store/app.store';
-import { IRepositoriesState } from '@/store/repositories.store';
+import { IAppState } from '@/store/aspects/app';
+import { IRepositoriesState } from '@/store/aspects/repositories';
 
 Vue.use(Vuex);
